+++ src/clientmetricreport/ClientMetricReport.ts
@@ -10,4 +10,14 @@ export default interface ClientMetricReport {
    * Gets raw client media metrics
    */
   getObservableMetrics(): { [id: string]: number };
+
+  /**
+   * Gets video upstream client media metrics
+   */
+  getObservableVideoUpstreamMetrics(): { [id: string]: {} };
+
+  /**
+   * Gets video downstream client media metrics
+   */
+  getObservableVideoDownstreamMetrics(): { [id: string]: {} };
 }