+++ test/connectionmonitor/SignalingAndMetricsConnectionMonitor.test.ts
@@ -103,6 +103,15 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
     videoUpstreamBitrate: RawMetrics = 100;
     availableSendBandwidth: RawMetrics = 100;
     availableReceiveBandwidth: RawMetrics = 100;
+    videoUpstreamPacketsSent: RawMetrics = 100;
+    videoUpstreamFramesEncodedPerSecond: RawMetrics = 100;
+    videoUpstreamFrameHeight: RawMetrics = 100;
+    videoUpstreamFrameWidth: RawMetrics = 100;
+    videoDownstreamBitrate: RawMetrics = 100;
+    videoDownstreamPacketLossPercent: RawMetrics = 100;
+    videoDownstreamFramesDecodedPerSecond: RawMetrics = 100;
+    videoDownstreamFrameHeight: RawMetrics = 100;
+    videoDownstreamFrameWidth: RawMetrics = 100;
 
     getObservableMetrics(): { [id: string]: number } {
       return {
@@ -112,6 +121,26 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
         videoUpstreamBitrate: this.videoUpstreamBitrate,
         availableSendBandwidth: this.availableSendBandwidth,
         availableReceiveBandwidth: this.availableReceiveBandwidth,
+      };
+    }
+
+    getObservableVideoUpstreamMetrics(): { [id: string]: {} } {
+      return {
+        videoUpstreamBitrate: this.videoUpstreamPacketsSent,
+        videoUpstreamPacketsSent: this.videoUpstreamPacketsSent,
+        videoUpstreamFramesEncodedPerSecond: this.videoUpstreamFramesEncodedPerSecond,
+        videoUpstreamFrameHeight: this.videoUpstreamFrameHeight,
+        videoUpstreamFrameWidth: this.videoUpstreamFrameWidth,
+      };
+    }
+
+    getObservableVideoDownstreamMetrics(): { [id: string]: {} } {
+      return {
+        videoDownstreamBitrate: this.videoDownstreamBitrate,
+        videoDownstreamPacketLossPercent: this.videoDownstreamPacketLossPercent,
+        videoDownstreamFramesDecodedPerSecond: this.videoDownstreamFramesDecodedPerSecond,
+        videoDownstreamFrameHeight: this.videoDownstreamFrameHeight,
+        videoDownstreamFrameWidth: this.videoDownstreamFrameWidth,
       };
     }
   }